Refactor item port auto push into self-contained class.
